Create ZerosInFact.cpp
Idea: For any N, N! can be expressed of the form (2^a) * (5^b) * M for some arbitrary non-zero M. It is trivial to see that, the amount of powers of 2 are more than powers of 5, therefore the limiting factor is the powers of 5. We must find all powers of 5 in the number A to find trailing zeros.