Create isPalindrome.cpp
Idea: Only consider half of the number. Find the amount of digits this number has. If it's odd, note it. Then, construct the half of this integer to a new variable while dividing the initial number. Compare at the end if they are equal or not. If A was odd, divide once more as the middle won't effect the palindrome state of the number.