Create RevInt.cpp
Idea: Since we work with an integer, we can pretend our result is a long long int. Then, as we add to this we will check if it overflows a lower limit, INT_MAX or the other way around with INT_MIN. In the end we will return the proper value casted automatically down to an int, since we know it's representable.